extend logic tester to account for more option combinations

+++ test-logic.sh
@@ -2,21 +2,26 @@
 # and produce an svg for visual inspection
 function store-svg() {
     local name=test-logic--"$1"; shift
+    local svg=${name}.svg
+    [[ -f $svg ]] && { echo $svg already present - skipped ; return 0; }
+    [[ -n "$DRY_RUN" ]] && { echo would update $svg from options "$@"; return 0; }
+    echo "==============================" RUNNING with options "$@"
     python -u ./demo-oai.py --devel -n -v "$@"
     rm demo-oai-graph.dot
-    mv demo-oai-graph.svg $name.svg
-    echo "==============================" $name from options "$@"
+    mv demo-oai-graph.svg $svg
+    echo "===============" DONE $svg from options "$@"
 }
 
 function tests-basic() {
-    store-svg "noopt"
-    store-svg "noopt-nok8reset" -k
-    store-svg "load" -l
-    store-svg "noauto" -a
-    store-svg "load-noauto" -l -a
-    store-svg "start" --start
-    store-svg "stop" --stop
-    store-svg "cleanup" --cleanup
+    store-svg "plain"
+    store-svg "plain-load" -l
+    store-svg "plain-noauto" -a
+    store-svg "plain-load-noauto" -l -a
+    store-svg "plain-nok8reset" -k
+    store-svg "plain-load-nok8reset" -l -k
+    store-svg "plain-start" --start
+    store-svg "plain-stop" --stop
+    store-svg "plain-cleanup" --cleanup
 }
 
 # with quectel selected
@@ -25,7 +30,13 @@ function tests-quectel() {
     store-svg "quectel2-load" -l -Q 9 -Q 18
     store-svg "quectel1" -Q 9
     store-svg "quectel2" -Q 9 -Q 18
+    store-svg "quectel1-noauto"  -Q 9 -a
+    store-svg "quectel1-cleanup" -Q 9 --start
+    store-svg "quectel1-cleanup" -Q 9 --stop
     store-svg "quectel1-cleanup" -Q 9 --cleanup
+    store-svg "quectel2-cleanup" -Q 9 -Q 18 --start
+    store-svg "quectel2-cleanup" -Q 9 -Q 18 --stop
+    store-svg "quectel2-cleanup" -Q 9 -Q 18 --cleanup
 }
 
 tests-basic
